components/host-components/hddp/server/requirements.txt
aniso8601==8.0.0
attrs==20.3.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
docker==4.4.0
Flask==1.1.2
Flask-Cors==3.0.9
flask-restx==0.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
pycparser==2.20
pyOpenSSL==19.1.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
requests==2.23.0
six==1.15.0
urllib3==1.25.11
websocket-client==0.57.0
Werkzeug==1.0.1